requirements.txt

pymongo>=4.0.0
bson>=0.5.10
aiohttp>=3.8.1
pytest>=7.0.0
pytest-asyncio>=0.18.0
python-dotenv>=0.19.0
pydantic>=1.9.0
typing-extensions>=4.0.0
requests>=2.27.0
beautifulsoup4>=4.10.0
lxml>=4.7.0
numpy>=1.21.0
pandas>=1.3.0
jsonschema>=4.0.0
python-dateutil>=2.8.2
pyyaml